+++ src/app/(new-version)/entity/[id]/page.tsx
@@ -2,28 +2,9 @@ import { MainContentWrapper } from '@/components-new-version/common/main-content
 import { ProfileContent } from '@/components-new-version/profile/profile-content'
 import { TokenContent } from '@/components-new-version/token/token-content'
 import {
-  isValidPublicKey,
-  isValidTransactionSignature,
-} from '@/components-new-version/utils/validation'
-
-export enum RouteType {
-  TRANSACTION = 'transaction',
-  TOKEN = 'token',
-  PROFILE = 'profile',
-}
-
-function renderContent(routeType: RouteType, cleanId: string) {
-  switch (routeType) {
-    case RouteType.TRANSACTION:
-      return <p>transaction page</p>
-    case RouteType.TOKEN:
-      return <TokenContent id={cleanId} />
-    case RouteType.PROFILE:
-      return <ProfileContent username={cleanId} />
-    default:
-      return <p>Unknown route type</p>
-  }
-}
+  determineRouteType,
+  RouteType,
+} from '@/components-new-version/utils/entity-page'
 
 export default async function Entity({ params }: { params: { id: string } }) {
   const { id } = params
@@ -31,21 +12,20 @@ export default async function Entity({ params }: { params: { id: string } }) {
   const cleanId = id.startsWith('@') ? id.slice(1) : id
   const routeType = determineRouteType(id)
 
+  function renderContent(routeType: RouteType, cleanId: string) {
+    switch (routeType) {
+      case RouteType.TRANSACTION:
+        return <p>transaction page</p>
+      case RouteType.TOKEN:
+        return <TokenContent id={cleanId} />
+      case RouteType.PROFILE:
+        return <ProfileContent username={cleanId} />
+      default:
+        return <p>Unknown route type</p>
+    }
+  }
+
   return (
     <MainContentWrapper>{renderContent(routeType, cleanId)}</MainContentWrapper>
   )
-}
-
-export function determineRouteType(id: string): RouteType {
-  const cleanId = id.startsWith('@') ? id.slice(1) : id
-
-  if (isValidTransactionSignature(cleanId)) {
-    return RouteType.TRANSACTION
-  }
-
-  if (isValidPublicKey(cleanId)) {
-    return RouteType.TOKEN
-  }
-
-  return RouteType.PROFILE
 }
